requirements.txt
altgraph==0.17.4
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
blinker==1.8.2
bottle==0.12.25
celery==5.4.0
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clr-loader==0.2.6
colorama==0.4.6
cryptography==43.0.0
Flask==3.0.0
flaskwebgui==1.1.3
gitdb==4.0.11
GitPython==3.1.40
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.1
MarkupSafe==2.1.5
packaging==24.1
pefile==2023.2.7
pillow==10.3.0
prompt_toolkit==3.0.47
proxy_tools==0.1.0
psutil==5.9.8
pycparser==2.22
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
python-dateutil==2.9.0.post0
pythonnet==3.0.3
pywebview==5.2
pywin32-ctypes==0.2.3
requests==2.32.2
setuptools==73.0.1
six==1.16.0
smmap==5.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
waitress==2.1.2
wcwidth==0.2.13
Werkzeug==3.0.3
